feat(demos): implements demo CRUD
Implements create method in the demos service and adds validation to the create DTO.  Updates the demos module to use the demo entity.  Adds global validation pipes to the application and updates the price column type in the demo table.

+++ src/demos/entities/demo.entity.ts
@@ -10,7 +10,7 @@ export class Demo {
   })
   title: string;
 
-  @Column('numeric', {
+  @Column('float', {
     default: 0,
   })
   price: number;
